perf: server-prefetch trip, receipts, bento list pages
Extends the leave pattern (#208) to the next batch of list pages: extract a
fetchX(supabase) shared by the client hook and a server prefetch, then
prefetch + dehydrate + HydrationBoundary in each page so first load paints
real rows from the HTML instead of waiting on a post-hydration client fetch.

trip / receipts / bento all have plain (non-user-scoped) list queries.
debt/approve (user-scoped, multi-query) and meetings (client page, needs a
server-component refactor) are tracked as follow-ups.

Closes #213
